Fixed an issue where internal state could be leaked through `this`
When executing user-provided computed functions from the queue,
`computedNotify` would call them like this:

    computedQueue[computedI]();

This issue is a sneaky one because I'm calling the functions using
syntax that is quite weird, but you can probably see what's happening
more clearly if written like this:

    computedQueue.12();

That's right! It's not a normal naked function call, but a method call!
This means that all computed functions get their `this` value set to
`computedQueue` and can do evil things like:

    computed(function() {
      this.push("hehe :3");
    });

The above snippet would corrupt `computedQueue` and cause a TypeError
to be thrown, oh no!
I've fixed the issue by calling the function indirectly, like this:

    (0, computedQueue[computedI])();

Using this even weirder syntax, we get the value of
`computedQueue[computedI]` then call that value, instead of calling the
method `[computedI]` on `computedQueue`. Hooray!